requirements.txt
# AI Agent Starter Requirements

# Core dependencies
openai>=1.0.0
anthropic>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Memory
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Tools
requests>=2.31.0
beautifulsoup4>=4.12.0

# Web UI (optional)
streamlit>=1.28.0
gradio>=4.0.0

# Development
pytest>=7.4.0
black>=23.0.0
mypy>=1.5.0
